pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "equipment-provisioner"
version = "0.1.0"
description = "Automated network device provisioning for Cambium, Mikrotik, Tarana, and Tachyon devices"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
keywords = ["network", "provisioning", "mikrotik", "cambium", "tarana", "tachyon", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "aiohttp>=3.9.0",
    "pyroute2>=0.7.0",
    "scapy>=2.5.0",
    "paramiko>=3.4.0",
    "asyncssh>=2.14.0",
    "librouteros>=3.2.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.5.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
gpio = ["OPi.GPIO>=0.5.0"]
snmp = ["pysnmp>=6.0.0"]
web = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = ["equipment-provisioner[gpio,snmp,web,dev]"]

[project.scripts]
provisioner = "provisioner.main:main"
provisioner-cli = "provisioner.cli:main"
provisioner-web = "provisioner.web_server:main"

[project.urls]
Homepage = "https://github.com/isolson/equipment-provisioner"
Documentation = "https://github.com/isolson/equipment-provisioner#readme"
Repository = "https://github.com/isolson/equipment-provisioner"
Issues = "https://github.com/isolson/equipment-provisioner/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["provisioner*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=provisioner --cov-report=term-missing"
